pyproject.toml
[tool.poetry]
name = "aiservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
langchain = "^0.3.1"
langchain-core = "^0.3.6"
langchain-community = "^0.3.1"
langchain-mistralai = "^0.2.0"
python-dotenv = "^1.0.1"
logger = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
